Road: Start implementation of roads

+++ core/src/main/java/me/rhin/openciv/game/unit/RangedUnit.java
@@ -43,7 +43,7 @@ public abstract class RangedUnit extends Unit
 		Civilization.getInstance().getEventManager().addListener(RelativeMouseMoveListener.class, this);
 	}
 
-	public abstract int getMovementCost(Tile prevTile, Tile adjTile);
+	public abstract float getMovementCost(Tile prevTile, Tile adjTile);
 
 	public abstract int getRangedCombatStrength();
