+++ core/src/main/java/me/rhin/openciv/game/unit/type/Archer.java
@@ -22,7 +22,7 @@ public class Archer extends UnitItem {
 		}
 
 		@Override
-		public int getMovementCost(Tile prevTile, Tile tile) {
+		public float getMovementCost(Tile prevTile, Tile tile) {
 			if (tile.containsTileProperty(TileProperty.WATER))
 				return 1000000;
 			else
@@ -44,7 +44,7 @@ public class Archer extends UnitItem {
 	public int getProductionCost() {
 		return 40;
 	}
-	
+
 	@Override
 	public float getGoldCost() {
 		return 100;
