+++ core/src/main/java/me/rhin/openciv/game/unit/type/Bowman.java
@@ -13,7 +13,7 @@ public class Bowman extends RangedUnit {
 	}
 
 	@Override
-	public int getMovementCost(Tile prevTile, Tile tile) {
+	public float getMovementCost(Tile prevTile, Tile tile) {
 		if (tile.containsTileProperty(TileProperty.WATER))
 			return 1000000;
 		else
