+++ core/src/main/java/me/rhin/openciv/game/unit/type/Galley.java
@@ -22,7 +22,7 @@ public class Galley extends UnitItem {
 		}
 
 		@Override
-		public int getMovementCost(Tile prevTile, Tile tile) {
+		public float getMovementCost(Tile prevTile, Tile tile) {
 			if (!tile.containsTileProperty(TileProperty.WATER))
 				return 1000000;
 			else
