+++ core/src/main/java/me/rhin/openciv/game/unit/type/Scout.java
@@ -21,7 +21,7 @@ public class Scout extends UnitItem {
 		}
 
 		@Override
-		public int getMovementCost(Tile prevTile, Tile tile) {
+		public float getMovementCost(Tile prevTile, Tile tile) {
 			if (tile.containsTileProperty(TileProperty.WATER))
 				return 1000000;
 			else if (tile.getMovementCost(prevTile) > 1 && tile.getMovementCost(prevTile) < 3)
@@ -31,7 +31,7 @@ public class Scout extends UnitItem {
 		}
 
 		@Override
-		public int getMaxMovement() {
+		public float getMaxMovement() {
 			return 3;
 		}
