+++ server/src/main/java/me/rhin/openciv/server/game/unit/type/Builder.java
@@ -45,7 +45,7 @@ public class Builder extends UnitItem {
 		}
 
 		@Override
-		public int getMovementCost(Tile prevTile, Tile tile) {
+		public float getMovementCost(Tile prevTile, Tile tile) {
 			if (tile.containsTileProperty(TileProperty.WATER))
 				return 1000000;
 			else
