+++ server/src/main/java/me/rhin/openciv/server/game/unit/type/Settler.java
@@ -20,7 +20,7 @@ public class Settler extends UnitItem {
 		}
 
 		@Override
-		public int getMovementCost(Tile prevTile, Tile tile) {
+		public float getMovementCost(Tile prevTile, Tile tile) {
 			if (tile.containsTileProperty(TileProperty.WATER))
 				return 1000000;
 			else
